fix warnings

+++ render/src/vulkan/descriptor_set.cpp
@@ -217,8 +217,8 @@ void update_bindless_set(Device &device, BindlessSet &bindless)
 	                           bindless.pending_unbind[BindlessSet::PER_IMAGE].size() +
 	                           bindless.pending_unbind[BindlessSet::PER_BUFFER].size();
 
-	auto *descriptor_writes       = tmp_scope.allocate<VkWriteDescriptorSet>(total_bind_count);
-	auto *descriptor_copies       = tmp_scope.allocate<VkCopyDescriptorSet>(total_unbind_count);
+	auto *descriptor_writes       = tmp_scope.allocate<VkWriteDescriptorSet>(u32(total_bind_count));
+	auto *descriptor_copies       = tmp_scope.allocate<VkCopyDescriptorSet>(u32(total_unbind_count));
 	auto *descriptor_writes_begin = descriptor_writes;
 	auto *descriptor_copies_begin = descriptor_copies;
 
@@ -228,9 +228,10 @@ void update_bindless_set(Device &device, BindlessSet &bindless)
 	usize total_buffer_info_count = bindless.pending_bind[BindlessSet::PER_BUFFER].size();
 
 	VkDescriptorImageInfo *image_infos =
-		total_image_info_count > 0 ? tmp_scope.allocate<VkDescriptorImageInfo>(total_image_info_count) : nullptr;
+		total_image_info_count > 0 ? tmp_scope.allocate<VkDescriptorImageInfo>(u32(total_image_info_count)) : nullptr;
 	VkDescriptorBufferInfo *buffer_infos =
-		total_buffer_info_count > 0 ? tmp_scope.allocate<VkDescriptorBufferInfo>(total_buffer_info_count) : nullptr;
+		total_buffer_info_count > 0 ? tmp_scope.allocate<VkDescriptorBufferInfo>(u32(total_buffer_info_count))
+									: nullptr;
 
 	VkDescriptorType descriptor_types[] = {
 		VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER,
